Clean state controls
