+++ front/src/pages/Main.js
@@ -6,9 +6,8 @@ import Searchbar from '../components/Searchbar'
 import MapContainer from '../components/MapContainer'
 import Results from '../components/Results'
 import { Layout } from 'antd'
-import Sider from 'antd/lib/layout/Sider'
 
-const { Header } = Layout;
+const { Header, Sider } = Layout;
 
 function Main(){
   const [keyword, setKeyword] = useState("");
@@ -19,33 +18,8 @@ function Main(){
   const [address3, setAddress3] = useState(null);
   const [tags, setTags] = useState([]);
 
-  const handleChange = (event) => {
-    setKeyword(event.target.value);
-  }
-  const changeAddress1 = (value) => {
-    setAddress1(value);
-    setAddress2(null);
-    setAddress3(null);
-  }
-  const changeAddress2 = (value) => {
-    setAddress2(value);
-    setAddress3(null);
-  }
-  const changeAddress3 = (value) => {
-    setAddress3(value);
-  }
-  const sliderChange = (value) => {
-    setPrice(value);
-  }
-  const rateChange = (value) => {
-    setRate(value);
-  }
-  const tagChange = (tag, checked) => {
-    const nextTags = checked ? [...tags, tag] : tags.filter(t => t !== tag);
-    setTags(nextTags);
-  }
-
-  const state = {keyword, price, rate, address1, address2, address3, tags};
+  const funcs = { setKeyword, setPrice, setRate, setAddress1, setAddress2, setAddress3, setTags }
+  const filterData = { keyword, price, rate, address1, address2, address3, tags }
 
   return (
     <Layout>
@@ -54,20 +28,11 @@ function Main(){
     </Header>
     <Layout hasSider={true}>
       <Sider>
-        <Searchbar
-          handleChange={handleChange}
-          sliderChange={sliderChange}
-          rateChange={rateChange}
-          tagChange={tagChange}
-          changeAddress1={changeAddress1}
-          changeAddress2={changeAddress2}
-          changeAddress3={changeAddress3}
-          state={state}
-        />
+        <Searchbar funcs={funcs} filterData={filterData} />
       </Sider>
       <Switch>
-        <Route path='/search' render={() => <Results state={state}/>} />
-        <Route path='/' render={() => <MapContainer state={state}/>} />
+        <Route path='/search' render={() => <Results filterData={filterData} />} />
+        <Route path='/' render={() => <MapContainer />} />
       </Switch>
     </Layout>
   </Layout>
